refactor(cli): split standard JSON module
